✨ ♻️ Use vote::result to close/finish a vote
A comment containing vote::result can be used to manually define a result (e.g. the result gathered at a meeting when electronic quorum is not met).

When the vote is closed in this way, the reason will be added to the Bot's vote summary, the vote/open label will be removed, and the vote/done label added in its place.